Make all the experimental subcommand consistent.
